+++ R/tag-registry.R
@@ -15,29 +15,11 @@ parse_tag <- function(x) {
 
 #' Register tag parsers.
 #'
+#' Use this if you're building your own roclet that uses new tags.
+#'
 #' @param ... List of name-parser pairs
-#' @return \code{NULL}
 #' @export
 #' @keywords internal
-#' @rdname register-parser
-#' @usage NULL
-register.preref.parser <- function(key, parser) {
-  warning("Deprecated: please use register_tag instead", call. = FALSE)
-  tags[[key]] <- parser
-}
-
-#' @export
-#' @rdname register-parser
-#' @usage NULL
-register.preref.parsers <- function(parser, ...) {
-  warning("Deprecated: please use register_tag instead", call. = FALSE)
-  for (key in c(...)) {
-    register.preref.parser(key, parser)
-  }
-}
-
-#' @export
-#' @rdname register-parser
 register_tags <- function(...) {
   list2env(list(...), tags)
 }